Big update, refactoring for multiple person development
- World 2 picker
- Updates to lots of code pieces!

+++ scripts/knockback_floppy/knockback_floppy.gml
@@ -1,12 +1,14 @@
 // Script assets have changed for v2.3.0 see
 
 	function knockback_floppy(){
-		if (oHero.knockingback == false){		
+		if(oHero.shrinking == false){
+			if (oHero.knockingback == false){		
 		
-			global.shake = true;
-			oHero.knockingback = true;
-			oHero.ySpeed -= 4;
-			oHero.alarm[2] = 10; //alarm to kill floppy in 3 seconds
+				global.shake = true;
+				oHero.knockingback = true;
+				oHero.ySpeed -= 4;
+				oHero.alarm[2] = 10; //alarm to kill floppy in 3 seconds
 
-	}
+			}
+		}
 }
